client_gRPC/requirements.txt
Flask==2.3.2
grpcio==1.67.1
grpcio-tools==1.67.1
protobuf==5.28.3